+++ ci/install-curl.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+# Install cURL
+
+case $CI_OS in
+  windows*)
+    choco install --no-progress curl
+    ;;
+esac
+
+# Print version
+curl --version