+++ ixmp/reporting/quantity.py
@@ -101,7 +101,7 @@ class AttrSeries(pd.Series):
                 # the level. Use .xs() to avoid this behaviour unless drop=True
                 return AttrSeries(self.xs(key, level=level, drop_level=False))
 
-        idx = tuple(indexers.get(l, slice(None)) for l in self.index.names)
+        idx = tuple(indexers.get(n, slice(None)) for n in self.index.names)
         return AttrSeries(self.loc[idx])
 
     def sum(self, *args, **kwargs):