+++ ixmp/tests/core/test_platform.py
@@ -138,8 +138,11 @@ def test_add_timeslice_duplicate_raise(test_mp):
 
 def test_weakref():
     """Weak references allow Platforms to be del'd while Scenarios live."""
-    mp = ixmp.Platform(backend='jdbc', driver='hsqldb',
-                       url=f'jdbc:hsqldb:mem:test_weakref')
+    mp = ixmp.Platform(
+        backend='jdbc',
+        driver='hsqldb',
+        url='jdbc:hsqldb:mem:test_weakref',
+    )
 
     # There is one reference to the Platform, and zero weak references
     assert getrefcount(mp) - 1 == 1